+++ lib/src/quran/presentation/widgets/download_fonts_page/quran_fonts_page.dart
@@ -32,6 +32,7 @@ class _QuranFontsPage extends StatelessWidget {
   final Widget? secondMenuChild;
   final void Function(AyahModel ayah)? secondMenuChildOnTap;
   final AyahLongClickStyle? ayahLongClickStyle;
+  final TafsirStyle? tafsirStyle;
 
   const _QuranFontsPage({
     required this.context,
@@ -61,6 +62,7 @@ class _QuranFontsPage extends StatelessWidget {
     this.secondMenuChild,
     this.secondMenuChildOnTap,
     this.ayahLongClickStyle,
+    this.tafsirStyle,
   });
 
   @override
@@ -89,34 +91,36 @@ class _QuranFontsPage extends StatelessWidget {
                             Responsive.isMobileLarge(context))
                     ? SingleChildScrollView(
                         child: PageBuild(
-                            pageIndex: pageIndex,
-                            surahNumber: surahNumber,
-                            bannerStyle: bannerStyle,
-                            isDark: isDark,
-                            surahNameStyle: surahNameStyle,
-                            surahInfoStyle: surahInfoStyle,
-                            onSurahBannerPress: onSurahBannerPress,
-                            basmalaStyle: basmalaStyle,
-                            textColor: textColor,
-                            bookmarks: bookmarks,
-                            onAyahLongPress: onAyahLongPress,
-                            secondMenuChild: secondMenuChild,
-                            secondMenuChildOnTap: secondMenuChildOnTap,
-                            bookmarkList: bookmarkList,
-                            ayahIconColor: ayahIconColor,
-                            showAyahBookmarkedIcon: showAyahBookmarkedIcon,
-                            bookmarksAyahs: bookmarksAyahs,
-                            bookmarksColor: bookmarksColor,
-                            ayahSelectedBackgroundColor:
-                                ayahSelectedBackgroundColor,
-                            isFontsLocal: isFontsLocal,
-                            fontsName: fontsName,
-                            ayahBookmarked: ayahBookmarked,
-                            anotherMenuChild: anotherMenuChild,
-                            anotherMenuChildOnTap: anotherMenuChildOnTap,
-                            context: context,
-                            quranCtrl: quranCtrl,
-                            ayahLongClickStyle: ayahLongClickStyle),
+                          pageIndex: pageIndex,
+                          surahNumber: surahNumber,
+                          bannerStyle: bannerStyle,
+                          isDark: isDark,
+                          surahNameStyle: surahNameStyle,
+                          surahInfoStyle: surahInfoStyle,
+                          onSurahBannerPress: onSurahBannerPress,
+                          basmalaStyle: basmalaStyle,
+                          textColor: textColor,
+                          bookmarks: bookmarks,
+                          onAyahLongPress: onAyahLongPress,
+                          secondMenuChild: secondMenuChild,
+                          secondMenuChildOnTap: secondMenuChildOnTap,
+                          bookmarkList: bookmarkList,
+                          ayahIconColor: ayahIconColor,
+                          showAyahBookmarkedIcon: showAyahBookmarkedIcon,
+                          bookmarksAyahs: bookmarksAyahs,
+                          bookmarksColor: bookmarksColor,
+                          ayahSelectedBackgroundColor:
+                              ayahSelectedBackgroundColor,
+                          isFontsLocal: isFontsLocal,
+                          fontsName: fontsName,
+                          ayahBookmarked: ayahBookmarked,
+                          anotherMenuChild: anotherMenuChild,
+                          anotherMenuChildOnTap: anotherMenuChildOnTap,
+                          context: context,
+                          quranCtrl: quranCtrl,
+                          ayahLongClickStyle: ayahLongClickStyle,
+                          tafsirStyle: tafsirStyle,
+                        ),
                       )
                     : PageBuild(
                         pageIndex: pageIndex,
@@ -146,7 +150,9 @@ class _QuranFontsPage extends StatelessWidget {
                         anotherMenuChildOnTap: anotherMenuChildOnTap,
                         context: context,
                         quranCtrl: quranCtrl,
-                        ayahLongClickStyle: ayahLongClickStyle)
+                        ayahLongClickStyle: ayahLongClickStyle,
+                        tafsirStyle: tafsirStyle,
+                      )
                 : PageBuild(
                     pageIndex: pageIndex,
                     surahNumber: surahNumber,
@@ -174,7 +180,9 @@ class _QuranFontsPage extends StatelessWidget {
                     anotherMenuChildOnTap: anotherMenuChildOnTap,
                     context: context,
                     quranCtrl: quranCtrl,
-                    ayahLongClickStyle: ayahLongClickStyle),
+                    ayahLongClickStyle: ayahLongClickStyle,
+                    tafsirStyle: tafsirStyle,
+                  ),
       ),
     );
   }