+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.4
 
 * **Add more optinos to AyahAudioStyle.**
+* **Add more optinos to TafsirStyle.**
 * **Add AyahLongClickStyle to change Ayah menu style.**
 * **Add IndexTabStyle to change Surah & Juz list style.**
 * **Add SearchTabStyle to change search tab style.**